Add Craco and move some files to other folders
