+++ craco.config.js
@@ -1,0 +1,15 @@
+const path = require("path");
+
+module.exports = {
+  webpack: {
+    alias: {
+      '@api': path.resolve(__dirname, 'src/api'),
+      '@app': path.resolve(__dirname, 'src/app'),
+      '@components': path.resolve(__dirname, 'src/components'),
+      '@entities': path.resolve(__dirname, 'src/entities'),
+      '@pages': path.resolve(__dirname, 'src/pages'),
+      '@store': path.resolve(__dirname, 'src/store'),
+      '@widgets': path.resolve(__dirname, 'src/widgets'),
+    }
+  }
+};
