+++ src/app/AppRouter.tsx
@@ -1,7 +1,7 @@
 import React, { FC, Suspense } from "react";
 import { Route, Routes } from "react-router-dom";
-import { Layout } from "./UI/Layout/Layout";
-import { LoginPage } from "../pages/LoginPage/LoginPage";
+import { Layout } from "@components/UI/Layout/Layout";
+import { LoginPage } from "@pages/LoginPage/LoginPage";
 
 const AppRouter: FC = () => {
   return (
